Switch to new PyO3 `Bound` API
This opts into the new PyO3 0.21 beta, which is scheduled for a release this month.

It also switcher to the new `Bound` API which should be more performant.

Also switches from using an `Iterator` for the `match_frame`s to a `List`, as I hope that will be slightly more performant.