feat: Add methods we'll need to build a django integration
The Pymemcache package allows us to inject a `hasher` which needs to
implement the protocol of get_node, add_node, remove_node. Adding those
methods to Ketama allows us to build a very simple wrapper around the
Rust binding that will work nicely with Django

Paired with @anonrig on this.