fix(dev): Try workaround to get maturin to include license file (#81)
Pypi expects us to declare the license file according to PEP 639[1], but maturin doesn't yet support it[2]. Another maturin user posted a workaround[3], to try to get maturin to include the license file in the tarball, which this PR implements.

(There is a draft PR[4] in the maturin repo to add support for PEP 639, but as of now it seems incomplete. If/when that's completed, merged, and released, we can update maturin and then fix the license specification as is done in https://github.com/getsentry/ophio/pull/80.)


[1] https://peps.python.org/pep-0639/
[2] https://github.com/PyO3/maturin/issues/2531#issuecomment-2833172453
[3] https://github.com/PyO3/maturin/issues/2531#issuecomment-2832632196
[4] https://github.com/PyO3/maturin/pull/2571